+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.1 &mdash; Coming Soon
+
+- Added hello-triangle example
+- Added PerspectiveCamera struct under the interactive feature
+- Make Sampler public (it should have been!)
+- Add texture-mapping example
+
 ## 0.1.0 &mdash; April 4th 2025
 
 - First public release